backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0 